Removed constant pool creation, started to fix literalpool usage.
Prevented creation of troublesome constant pools by reducing return value
from a region to a 32-bit value. 0 indicates return, negative indicates exception,
positive indicates exit.

Started removal of all references to a LiteralPool in assembly code so that each region's
literal pool can be moved to the front of the region, so the only necessary reference to
it can always reach it.